+++ src/goatway/utils/xmpp.clj
@@ -1,6 +1,7 @@
 (ns goatway.utils.xmpp
   (:require [clojure.tools.logging :as log]
-            [goatway.runtime.db :as db])
+            [goatway.runtime.db :as db]
+            [goatway.channels.xmpp.transformer :as xmpp-transformer])
   (:import (org.jivesoftware.smackx.muc MultiUserChat)
            (org.jivesoftware.smack XMPPException$XMPPErrorException XMPPException$StreamErrorException AbstractConnectionListener)
            (org.jivesoftware.smack.packet XMPPError$Condition)))
@@ -15,12 +16,21 @@
             (.join muc try-nick)
             (dosync (ref-set last-err nil))
             (catch XMPPException$XMPPErrorException e
-              (if (= XMPPError$Condition/conflict (.getCondition (.getXMPPError e)))
-                (log/warn (str "Nickname " try-nick " conflicts, trying another..."))
-                (log/error (str "Unknown error " e)))
-              (dosync (ref-set last-err e))))
+              (dosync
+                (if (= XMPPError$Condition/conflict (.getCondition (.getXMPPError e)))
+                  (ref-set last-err :conflict)
+                  (if (= XMPPError$Condition/jid_malformed (.getCondition (.getXMPPError e)))
+                    (ref-set last-err :jid_malformed)
+                    (ref-set last-err e))))))
           (if @last-err
-            (recur (dec attempt) last-err (str nick "-" (- 10 attempt)))
+            (case @last-err
+              :conflict
+              (do (log/warn (str "Nickname " try-nick " conflicts, trying another..."))
+                  (recur (dec attempt) last-err (str nick "-" (- 10 attempt))))
+              :jid_malformed
+              (do (log/warn (str "Nickname " try-nick " is invalid for xmpp, trying another..."))
+                  (recur (dec attempt) last-err (str "anonymous-" (- 10 attempt))))
+              (log/error (str "Unknown error " @last-err)))
             {:nick try-nick})))))
 
 (defn create-listener
@@ -34,9 +44,11 @@
              (log/error error)
              (do (log/infof "smack: joined muc as %s" as)
                  (reset! nick as)
-                 (swap! db/puppets conj as)))))
+                 (swap! db/puppets conj as)
+                 (swap! xmpp-transformer/puppets-to-tg-users assoc as sender)))))
        (connectionClosedOnError [e]
          (swap! db/puppets disj @nick)
+         (swap! xmpp-transformer/puppets-to-tg-users dissoc @nick)
          (if (and
                (instance? XMPPException$StreamErrorException e)
                (= XMPPError$Condition/conflict (-> ^XMPPException$StreamErrorException e
@@ -45,4 +57,4 @@
            (log/debug e))
          (if error-callback (error-callback))))))
   ([muc sender]
-    (create-listener muc sender nil)))
+   (create-listener muc sender nil)))